:arrow_heading_up: Moved home, dashboard and login views to home_view.py & updated links
