+++ app.py
@@ -1,23 +1,23 @@
 #!/usr/bin/python3
 
-from collections import defaultdict
-from datetime import timedelta
-from flask import (Flask, request, render_template,
-                   session, redirect, url_for, escape)
 import glob
 import json
 import os
-import requests
 import uuid
+from datetime import timedelta
+
+from flask import (Flask, request, render_template,
+                   session, redirect, url_for, escape)
 from waitress import serve
 
-from bin.view_modifiers import response
 from bin import logger
+from bin.view_modifiers import response
+from views.home_view import load_home
 
 # Flask logging
 logthis = logger.setup_child_logger('jawa', 'app')
 error_message = ""
-verify_ssl = True  # Enables Jamf Pro SSL certificate verification
+
 
 # Initiate Flask
 app = Flask(__name__)
@@ -63,7 +63,7 @@ def register_blueprints():
     from views.okta_webhook import blueprint
     app.register_blueprint(blueprint)
     # Create a new Cron Job
-    from views.cron_views import blueprint
+    from views.cron_view import blueprint
     app.register_blueprint(blueprint)
     # Log view
     from views import log_view
@@ -77,6 +77,9 @@ def register_blueprints():
     # Webhooks Base view
     from views import webhook_view
     app.register_blueprint(webhook_view.blueprint)
+    # Home, Dashboard and Login view
+    from views import home_view
+    app.register_blueprint(home_view.blueprint)
 
 
 # Server setup including making .json file necessary for webhooks
@@ -85,7 +88,7 @@ def register_blueprints():
 @app.route('/setup', methods=['GET', 'POST'])
 def setup():
     if 'username' not in session:
-        return redirect(url_for('logout', error_title="Session Timed Out", error_message="Please sign in again"))
+        return redirect(url_for('home_view.logout', error_title="Session Timed Out", error_message="Please sign in again"))
     if request.method == 'POST':
         logthis.debug(f"[{session.get('url')}] {session.get('username')} /setup - POST")
         server_url = request.form.get('address')
@@ -102,10 +105,8 @@ def setup():
         new_json = {}
         if server_url != '':
             new_json['jawa_address'] = server_url
-        # print(server_url)
         if jps_url != '':
             new_json['jps_url'] = jps_url
-        # print(new_json)
         if not os.path.isfile(server_json_file):
             with open(server_json_file, "w") as outfile:
                 server_json = {'jawa_address': server_url, 'jps_url': jps_url, 'alternate_jps': jps_url2}
@@ -147,7 +148,7 @@ def setup():
 @response(template_file="setup/cleanup.html")
 def cleanup():
     if 'username' not in session:
-        return redirect(url_for('logout', error_title="Session Timed Out", error_message="Please sign in again"))
+        return redirect(url_for('home_view.logout', error_title="Session Timed Out", error_message="Please sign in again"))
     if request.method != 'POST':
         return {"username": session.get('username'), "scripts_dir": scripts_directory}
     logthis.info(f"[{session.get('url')}] {session.get('username')} is cleaning up scripts...")
@@ -162,179 +163,16 @@ def cleanup():
     return redirect(url_for('success'))
 
 
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    if request.method == 'POST':
-        if os.path.isfile(server_json_file):
-            with open(server_json_file) as json_file:
-                server_json = json.load(json_file)
-            if request.form.get('active_url'):
-                session['url'] = str(request.form.get('active_url'))
-            elif server_json.get('jps_url', 0):
-                if server_json['jps_url'] is not None and len(server_json['jps_url']) != 0:
-                    session['url'] = str(server_json['jps_url'])
-            elif request.form.get('url')[-1:] == "/":
-                session['url'] = str(request.form.get('url')).rstrip(request.form.get('url')[-1])
-            else:
-                session['url'] = request.form['url']
-
-        elif request.form.get('active_url')[-1:] == "/":
-            session['url'] = str(request.form.get('url')).rstrip(request.form.get('url')[-1])
-        else:
-            session['url'] = request.form['url']
-        session['username'] = request.form['username']
-        session['password'] = request.form['password']
-
-        logthis.info(f"[{session.get('url')}] Attempting login for: {session.get('username')}")
-
-        if request.form['password'] == "":
-            title = "Authentication error"
-            msg = "Passwords can't be blank"
-            return redirect(url_for('logout', error_title=title, error_message=msg))
-        try:
-            resp = requests.get(
-                session['url'] + '/JSSResource/activationcode',
-                auth=(session['username'], session['password']),
-                headers={'Accept': 'application/json'},
-                verify=verify_ssl)
-
-            resp.raise_for_status()
-
-        except requests.exceptions.HTTPError as err:
-            logthis.info(f"Error occurred: {err}")
-            return redirect(url_for('logout', error_title="HTTP Error", error_message=err))
-        except requests.exceptions.ConnectTimeout as err:
-            logthis.info(f"Error occurred: {err}")
-            return redirect(url_for('logout', error_title="Connection Timeout", error_message=err))
-        except requests.exceptions.ConnectionError as err:
-            logthis.info(f"Error occurred: {err}")
-            return redirect(url_for('logout', error_title="HTTP Error", error_message=err))
-
-        logthis.info(
-            f"[{session.get('url')}] Logging In: " + str(escape(session['username'])))
-
-        return redirect(url_for('dashboard'))
-
-    if 'username' not in session:
-        return redirect(url_for('logout', error_title="Session Timed Out", error_message="Please sign in again"))
-
-    return redirect(url_for('dashboard'))
-
-
-@app.route('/home.html')
-def index():
-    return load_home()
-
-
-def load_home(error_title="", error_message=""):
-    if 'username' in session:
-        return redirect(url_for('dashboard'))
-    if not os.path.isfile(server_json_file):
-        with open(server_json_file, "w") as fout:
-            json.dump({}, fout)
-
-    with open(server_json_file, "r") as fin:
-        server_json = json.load(fin)
-    if not server_json:
-        return render_template('home.html')
-    brand = server_json.get("brand")
-
-    if (
-            'jps_url' not in server_json
-            or server_json['jps_url'] is None
-            or len(server_json['jps_url']) == 0
-    ):
-        return render_template('home.html', app_name=brand)
-    if 'alternate_jps' not in server_json:
-        return render_template('home.html', app_name=brand, error_title=error_title, error_message=error_message)
-
-    if server_json['alternate_jps'] != "":
-        return render_template('home.html',
-                               jps_url=server_json['jps_url'],
-                               jps_url2=server_json['alternate_jps'],
-                               welcome="true", jsslock="true", app_name=brand, error_title=error_title,
-                               error_message=error_message)
-
-    session['url'] = server_json['jps_url']
-    return render_template('home.html',
-                           jps_url=str(escape(session['url'])),
-                           welcome="true", jsslock="true", app_name=brand, error_title=error_title,
-                           error_message=error_message)
-
-
 @app.route("/")
 def home():
     return load_home()
-
-
-@app.route("/dashboard")
-def dashboard():
-    if 'username' not in session:
-        return redirect(url_for('logout', error_title="Session Timed Out", error_message="Please sign in again"))
-    logthis.info(f"[{session.get('url')}] {session.get('username')} rendering /dashboard.")
-    with open(webhooks_file) as webhook_json:
-        webhooks_installed = json.load(webhook_json)
-    jamf_pro_webhooks = []
-    okta_webhooks = []
-    custom_webhooks = []
-    for each_webhook in webhooks_installed:
-        data = defaultdict(lambda: "MISSING", each_webhook)
-        tag = data['tag']
-        if tag == "jamfpro":
-            jamf_pro_webhooks.append(each_webhook)
-        elif tag == "okta":
-            okta_webhooks.append(each_webhook)
-        elif tag == "custom":
-            custom_webhooks.append(each_webhook)
-
-    data = []
-
-    if not os.path.isfile(cron_file):
-        with open(cron_file, "w") as outfile:
-            json.dump(data, outfile)
-
-    with open(cron_file, "r") as cron_json:
-        try:
-            cron_list = json.load(cron_json)
-        except json.decoder.JSONDecodeError as err:
-            with open(cron_file, "w") as cron_json:
-                cron_list = []
-                json.dump(cron_list, cron_json, indent=4)
-
-    cron_json = []
-    for cron in cron_list:
-        script = cron['script'].rsplit('/', 1)
-        cron_json.append({"name": cron['name'],
-                          "frequency": cron['frequency'],
-                          "script": script[1],
-                          "description": cron['description']})
-
-    webhook_url = session['url']
-
-    if not cron_json:
-        cron_json = ''
-    logthis.info(f"Total webhooks managed by JAWA: {len(webhooks_installed)}")
-    if webhook_json == cron_json:
-        return redirect(url_for('first_automation'))
-    return render_template(
-        'dashboard.html',
-        webhook_url=webhook_url,
-        jamfpro_list=jamf_pro_webhooks,
-        url=session.get('url'),
-        cron_list=cron_json,
-        okta_list=okta_webhooks,
-        custom_list=custom_webhooks,
-        total_webhooks=len(webhooks_installed),
-        total_cron=len(cron_json),
-        login="true",
-        username=str(escape(session['username'])))
 
 
 @app.route('/success', methods=['GET', 'POST'])
 def success():
     if 'username' not in session:
         logthis.info("No user logged in - returning to login page.")
-        return redirect(url_for('logout', error_title="Session Timed Out", error_message="Please sign in again"))
+        return redirect(url_for('home_view.logout', error_title="Session Timed Out", error_message="Please sign in again"))
     return render_template(
         'success.html',
         success_msg="",
@@ -347,7 +185,7 @@ def error():
     error_title = request.args.get('error')
     error_message = request.args.get('error_message')
     if 'username' not in session:
-        return redirect(url_for('logout'))
+        return redirect(url_for('home_view.logout'))
     logthis.info(
         f"[{session.get('url')}] {session.get('username').title()} was a victim of a series of accidents, as are we all. (/error)")
     return render_template('error.html', username=session.get('username'), error_message=error_title,
@@ -359,20 +197,10 @@ def page_not_found(error):
     if 'username' in session:
         logthis.info(
             f"[{session.get('url')}] {session.get('username')} wandered off course  ({request.path}) - redirecting to /dashboard.")
-        return redirect(url_for('dashboard'))
+        return redirect(url_for('home_view.dashboard'))
     logthis.info(
         f"An invalid path ({request.path}) was provided and no user is logged in.  Returning login page.")
     return load_home()
-
-
-@app.route('/logout')
-def logout():
-    error_title = request.args.get('error_title')
-    error_message = request.args.get('error_message')
-    if session.get('username'):
-        logthis.info("Logging Out: " + str(escape(session['username'])))
-        session.pop('username', None)
-    return load_home(error_title, error_message)
 
 
 if __name__ == '__main__':
